Add files via upload
dta_pcoef_0.5.gs is a GrADs script for generating the patten-change kernels (PCKs) of precipitation based on a set of CMIP6 models. 
gen_multimodel_dcxy.gs is a GrADs script for creating multi-model mean PCK based on a set of individual model PCK.
gen_basecov_monthly_0.5.f90 is a fortran code for performing spatial diaggreation-bias correction procedure to create a large ensemble global dataset, using the basecov scenario as an example.